Don't let hash function output leak into deadness analysis' output
In deadness analysis we try to canonicalize and and or expressions by sorting
them (so that e.g. A&B and B&A map ultimately map to the same Predicate*).  We
used to sort them by their hashes (this was a recursive "content" hash and did
not depend on pointer values) but this meant that changes to the hash function
would result in changes to the Predicates we'd produce for a TensorFlow node.

This CL avoids this unnecessary dependence on hashes by using a deterministic ID
(order of allocation, basically) instead of the hash to sort the predicates.

PiperOrigin-RevId: 232372151