Added support for dequantization of inputs for some operators.
When an operator's input is floating-point but the weights or bias are quantized, Dequantize operators are added to convert quantized inputs to floating-point since NNAPI does not support mixing float and quantized types. This is needed for models using post-training quantization.

Only CONV_2D and FULLY_CONNECTED operations support this for now.

PiperOrigin-RevId: 232699006